requirements.txt
# Telegram Bot 框架
python-telegram-bot==20.7

# AI 服务
openai==1.3.7

# Web 控制面板
Flask==3.0.0
Flask-CORS==4.0.0

# 配置和数据处理
python-dotenv==1.0.0
requests==2.31.0

# 异步支持
asyncio==3.4.3
aiohttp==3.9.1

# 日志
loguru==0.7.2

# 定时任务
APScheduler==3.10.4

# JSON 处理
ujson==5.8.0